feat: change styles, minor refactor
- add adaptive styles
- refactor: moved out Header, ErrorButton into separate components
