+++ src/components/Details/Details.module.scss
@@ -8,7 +8,6 @@
   align-content: space-around;
   border: 1px solid #fff;
   border-radius: 8px;
-  flex: 0 1 50%;
   padding: 8px;
 
   .button {
