+++ src/components/Header/Header.module.scss
@@ -1,0 +1,3 @@
+.header {
+  padding: 2rem 2rem 0 2rem;
+}
