+++ src/components/List/List.module.scss
@@ -1,7 +1,5 @@
 .list {
-  display: flex;
-  flex-wrap: wrap;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
   gap: 16px;
-  flex: 1;
-  align-content: flex-start;
 }
