+++ src/components/index.ts
@@ -1,6 +1,8 @@
 export * from './Details';
 export * from './Error';
 export * from './ErrorBoundary';
+export * from './ErrorButton';
+export * from './Header';
 export * from './List';
-export * from './Search';
 export * from './Pagination';
+export * from './Search';
