+++ src/pages/Home/Home.module.scss
@@ -1,34 +1,13 @@
 .home {
-  padding: 2rem;
-
-  .button {
-    border-radius: 8px;
-    border: 1px solid transparent;
-    padding: 0.6em 1.2em;
-    font-size: 1em;
-    font-weight: 500;
-    font-family: inherit;
-    background-color: var(--error-color);
-    cursor: pointer;
-    transition:
-      box-shadow 0.25s,
-      filter 0.25s ease,
-      transform 0.05s;
-
-    &:hover {
-      box-shadow: 0 0 15px var(--box-shadow-color);
-      filter: brightness(1.2);
-    }
-
-    &:active {
-      transform: scale(0.98);
-    }
-  }
+  min-height: 100vh;
+  min-height: 100dvh;
+  display: grid;
+  grid-template-rows: auto 1fr auto;
 
   .wrapper {
-    display: flex;
+    padding: 2rem 2rem 0 2rem;
+    display: grid;
     gap: 16px;
-    justify-content: space-between;
-    margin-top: 32px;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
   }
 }
