+++ src/pages/Home/Home.tsx
@@ -1,7 +1,6 @@
 import {
   useState,
   useEffect,
-  ChangeEvent,
   useCallback,
   useRef,
   Dispatch,
@@ -10,7 +9,7 @@ import {
 import { Outlet } from 'react-router';
 import { useSearchParams } from 'react-router-dom';
 import { Character, TResponse } from '~/types';
-import { List, Search, Pagination } from '~/components';
+import { Header, List, Pagination } from '~/components';
 import { useLocalStorage, LSKey } from '~/hooks';
 import styles from './Home.module.scss';
 
@@ -24,7 +23,6 @@ export const BASE_URL = 'https://swapi.dev/api/people/';
 export const Home = () => {
   const [characters, setCharacters] = useState<Character[]>([]);
   const [isLoading, setIsLoading] = useState(false);
-  const [hasError, setHasError] = useState(false);
   const [totalCount, setTotalCount] = useState(0);
   const [searchParams, setSearchParams] = useSearchParams();
   const [currentPage, setCurrentPage] = useState(
@@ -83,25 +81,9 @@ export const Home = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [currentPage]);
 
-  useEffect(() => {
-    if (hasError) throw new Error('Your bad =(');
-  }, [hasError]);
-
-  const searchInputHandler = (e: ChangeEvent<HTMLInputElement>) =>
-    setSearchValue(e.target.value);
-
-  const throwError = () => setHasError(true);
-
   return (
     <div className={styles.home}>
-      <button className={styles.button} onClick={throwError}>
-        Generate error
-      </button>
-      <Search
-        value={searchValue}
-        onChange={searchInputHandler}
-        onSubmit={fetchCharacters}
-      />
+      <Header fetchCharacters={fetchCharacters} />
       <div className={styles.wrapper} ref={wrapperRef}>
         {isLoading ? (
           <h2 style={{ margin: 'auto' }}>Loading...</h2>
